Uncommented User REST, Added delete safety

+++ src/main/java/com/example/demo/services/FriendService.java
@@ -2,16 +2,23 @@ package com.example.demo.services;
 
 
 import com.example.demo.entities.Friend;
+import com.example.demo.entities.User;
 import com.example.demo.repositories.FriendRepo;
+import com.example.demo.repositories.UserRepo;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class FriendService {
     @Autowired
     private FriendRepo friendRepo;
+
+    @Autowired
+    private UserRepo userRepo;
 
     public List<Friend> findById(long user_id) {
         if(friendRepo.findByUserid(user_id) != null){
@@ -30,7 +37,25 @@ public class FriendService {
         return null;
     }
 
+    public Boolean checkIfUser(Long id){
+        UserService userService = new UserService();
+        Optional<Friend> friendList = friendRepo.findById(id);
+
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        User user = userRepo.findByUsername(username);
+
+        if(user.getId() == friendList.get().getUser_id())
+            return true;
+        else
+            return false;
+    }
+
     public void deleteById(long id) {
-        friendRepo.deleteById(id);
+        if(checkIfUser(id)) {
+            System.out.println("Deleting friendship " + id);
+            friendRepo.deleteById(id);
+        }
+        else
+            System.err.println("Cannot, simply can´t");
     }
 }
